requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# ONNX export & inference
onnx>=1.14.0
onnxruntime>=1.15.0

# Training utilities
tqdm>=4.65.0
matplotlib>=3.7.0

# API server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Knowledge distillation (optional)
# No additional deps beyond PyTorch
